+++ example/lib/pages/network_event_page.dart
@@ -31,6 +31,7 @@ class _NetworkEventPageState extends State<NetworkEventPage> {
         if (_events.length > 8) _events.removeLast();
       });
     });
+    unawaited(_loadCurrentNetwork());
   }
 
   @override